requirements.txt
streamlit==1.31.0
pandas==2.1.4
selenium==4.16.0
webdriver-manager==4.0.1
psycopg2-binary==2.9.9
matplotlib==3.8.2
numpy==1.26.3
altair==5.2.0
python-dotenv==1.0.0